.github/skills/experimental/tts-voiceover/pyproject.toml
[project]
name = "tts-voiceover-skill"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
    "azure-cognitiveservices-speech>=1.41",
    "azure-identity>=1.19",
    "python-pptx>=1.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "ruff>=0.15",
]
fuzz = [
    "atheris>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]
python_files = ["test_*.py", "fuzz_harness.py"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
